Reduction of harmony overhead
Uses Prepare() to gate  if harmony patches should apply or not. Also removes some reflection, switching to the use of publicizer instead.

Note that one of the tradeoffs is that the game will need a restart to change mod options. The use of a patch ledger could allow for the best of both worlds, if deemed necessary.